restapi_sorce/requirements.txt
mysqlclient-1.4.6-cp37-cp37m-win32.whl
pymysql==1.0.2
aiohttp==3.6.2
aiohttp-security==0.4.0
aiohttp-session==2.9.0
aiohttp-swagger==1.0.16
aioredis==2.0.1
alembic==1.4.2
altgraph==0.17.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
baidu-aip==2.2.17.0
blinker==1.4
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
colorama==0.4.3
connexion[swagger-ui]
cryptography==2.9.2
cx-Freeze==6.10
cx-Logging==3.0
Cython==0.29.21
Deprecated==1.2.13
Flask==2.0.3
Flask-Babel==1.0.0
Flask-Cors==3.0.8
Flask-Excel==0.0.7
Flask-Migrate==2.5.3
Flask-QRcode==3.0.0
flask-redis==0.4.0
Flask-Redis-Sentinel==2.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
geojson==2.5.0
googletrans==3.1.0a0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.25
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
inflection==0.5.0
iso8601==0.1.16
itsdangerous==2.1.0
Jinja2==3.0.3
Js2Py==0.70
jsonschema==3.2.0
kaitaistruct==0.8
ldap3==2.7
lief==0.12.1
lml==0.1.0
Mako==1.1.3
MarkupSafe==2.1.0
numpy==1.19.5
openapi-spec-validator==0.2.9
packaging==21.3
paho-mqtt==1.5.1
passlib==1.7.2
pefile==2021.9.3
Pillow==6.1.0
protobuf==3.11.3
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyjsparser==2.7.1
PyJWT==1.7.1
pymssql==2.1.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.16.0
pyserial==3.5
python-consul==1.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qrcode==6.1
redis==4.1.4
Redis-Sentinel-Url==1.0.1
requests==2.24.0
rfc3986==1.4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sniffio==1.1.0
sortedcontainers==2.1.0
speaklater==1.3
SQLAlchemy==1.4.31
swagger-spec-validator==2.7.3
swagger-ui-bundle==0.0.8
texttable==1.6.3
tornado==6.0.4
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.10
urwid==2.0.1
Werkzeug==2.0.3
wrapt==1.13.3
wsproto==0.15.0
xlrd==1.2.0
XlsxWriter==1.4.2
xlwt==1.3.0
yarl==1.4.2
zipp==3.1.0
zstandard==0.13.0
